Fix warnings reported by Clang
* Fix link options to work with Clang
* Fix Clang warnings around subobject initialization
* Avoid instantiation after specialization
* Clarify intention with hidden overloads
* Use correct cast for printf format specifier
* Avoid empty struct compatibility issues
* Support clang function naming in unit tests

+++ tools/legacy/sample_vpp/src/sample_vpp_utils.cpp
@@ -253,7 +253,7 @@ void PrintInfo(sInputParams* pParams, mfxVideoParam* pMfxParams, MFXVideoSession
     msdk_printf(MSDK_STRING("IOpattern type               \t%s\n"),
                 IOpattern2Str(pParams->IOPattern));
     msdk_printf(MSDK_STRING("Number of asynchronious tasks\t%hu\n"),
-                (unsigned int)pParams->asyncNum);
+                (unsigned short)pParams->asyncNum);
     msdk_printf(MSDK_STRING("Time stamps checking         \t%s\n"),
                 pParams->ptsCheck ? MSDK_STRING("ON") : MSDK_STRING("OFF"));
